+++ src/main/java/com/github/oleksii/zinkevych/currency_exchange_point/model/request/ApplicationCreateRequest.java
@@ -1,5 +1,6 @@
 package com.github.oleksii.zinkevych.currency_exchange_point.model.request;
 
+import java.math.BigDecimal;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
@@ -11,6 +12,7 @@ import lombok.Data;
 public class ApplicationCreateRequest {
 
     private final static int CURRENCY_NAME_LENGTH = 3;
+
     private final static int NAME_MAX_LENGTH = 50;
 
     @NotBlank(message = "The 'saleCurrency' must be specified")
@@ -24,7 +26,7 @@ public class ApplicationCreateRequest {
     private String purchaseCurrency;
 
     @NotNull(message = "The 'purchaseAmount' must be specified")
-    private Double purchaseAmount;
+    private BigDecimal purchaseAmount;
 
     @NotBlank(message = "The 'firstName' must be specified")
     @Size(max = NAME_MAX_LENGTH, message = "Length of 'firstName' must be less or equals {max} characters")
@@ -35,7 +37,7 @@ public class ApplicationCreateRequest {
     private String lastName;
 
     @NotBlank(message = "The 'phone' must be specified")
-    @Pattern(regexp="^(\\+)?([- _():=+]?\\d[- _():=+]?){10,13}(\\s*)?$")
-    @Size(min=9,max=13)
+    @Pattern(regexp = "^(\\+)?([- _():=+]?\\d[- _():=+]?){10,13}(\\s*)?$")
+    @Size(min = 9, max = 13)
     private String phone;
 }
